+++ batch-console-api/src/main/java/io/github/pinpols/batch/console/domain/rbac/support/captcha/AliyunCaptchaVerifier.java
@@ -7,6 +7,7 @@ import io.github.pinpols.batch.common.http.OutboundHttpRequest;
 import io.github.pinpols.batch.common.http.OutboundHttpResponse;
 import io.github.pinpols.batch.common.http.OutboundHttpTransport;
 import io.github.pinpols.batch.console.config.CaptchaProperties;
+import io.github.pinpols.batch.console.support.http.ConsoleOutboundTransport;
 import java.io.IOException;
 import java.security.GeneralSecurityException;
 import java.time.Duration;
@@ -55,7 +56,7 @@ public class AliyunCaptchaVerifier implements CaptchaVerifier {
   public AliyunCaptchaVerifier(
       CaptchaProperties properties,
       ObjectMapper objectMapper,
-      OutboundHttpTransport httpTransport) {
+      @ConsoleOutboundTransport OutboundHttpTransport httpTransport) {
     this.properties = properties;
     this.objectMapper = objectMapper;
     this.httpTransport = httpTransport;